Enable no-any-return-explicit, and correct the plotnltepops arguments
pyrefly: of the 115 error kinds that the configuration does not name,
107 give no error, because preset = "strict" already reports them. Eight
more kinds fire. Only no-any-return-explicit is worth the change: it
found one function that declared float and returned the Any that
DataFrame.item() gives. Wrap that value in float() and enable the kind.

The other seven stay off. unused-call-result (858), unknown-argument-type
(619), explicit-any (352), implicit-bool (342), and unknown-variable-type
(246) each report a normal pattern of this code. unbound-name (82) stays
off, as before. unused-type-ignore (4) reports the comments that mypy
needs, e.g. "# type: ignore[arg-type]", because pyrefly gives no such
error, thus it cannot be enabled while refurb runs mypy.

plotnltepops selected model grid cell 0 without a message, thus a command
that gave a time but no cell reported that the file nlte_0000.out does
not exist. It now selects cell 0 only for a model of one cell, and asks
for -modelgridindex for a larger model. It also gives the number of cells
in the model, and refuses a cell outside that range. The error that asks
for a time now says that a cell can also be needed.

Write the units of an axis label in square brackets, which is the
convention of the other commands. Six labels used parentheses. Two labels
of plotnltepops named the same quantity, thus give both the name that the
time and velocity plot already uses.

+++ artistools/transitions.py
@@ -246,7 +246,7 @@ def make_plot(
 
 def get_lte_partfunc(pldflevels: pl.DataFrame, T_exc: float) -> float:
     """Return the LTE partition function of the ion at the excitation temperature."""
-    return pldflevels.select(pl.col("g") * (-pl.col("energy_ev") / K_B_ev_per_K / T_exc).exp()).sum().item()
+    return float(pldflevels.select(pl.col("g") * (-pl.col("energy_ev") / K_B_ev_per_K / T_exc).exp()).sum().item())
 
 
 def add_upper_lte_pop(
